Version 0.6 (#8)
* Better reboot

It is better to use svc power to reboot through the power manager instead of forcing reboot. I think it prevents corruption of data. If it fails - it will force reboot like before.

* Add checking for bootctl utility

This checking for bootctl utility prevents the crash of the app when the utility is not available like in MagiskSU (without the module that fixes that).
It shows a dialog that informs the user about that.

* Change to version 0.6

Changing the version to v0.6

* Add screenshot file of v0.6

Screenshot of v0.6

* Update README.md

Added instructions for MagiskSU and updated the screenshot to v0.6

* Remove unnecessary imports