Version 0.6-rc2 (#10)
* Better reboot

It is better to use svc power to reboot through the power manager instead of forcing reboot. I think it prevents corruption of data. If it fails - it will force reboot like before.

* Add checking for bootctl utility

This checking for bootctl utility prevents the crash of the app when the utility is not available like in MagiskSU (without the module that fixes that).
It shows a dialog that informs the user about that.

* Change to version 0.6

Changing the version to v0.6

* Add screenshot file of v0.6

Screenshot of v0.6

* Update README.md

Added instructions for MagiskSU and updated the screenshot to v0.6

* Remove unnecessary imports

Removed unnecessary imports.

* Changed the reading of the buffer

The readLine function with DataInputStream is depracted - switching to BufferedReader.
Also removing the call to dismiss the error dialog because the dialog is already dismissed.

* Update README.md

Removed ARM64 requirement.

* Fix detection of A/B device

Fixed the detection of A/B device and added displaying of error dialog if android version is lower than Nougat 7.0

* Update README.md

Added requirement for android Nougat 7.0 or newer.

* Correcting minimum API level

Minimum android version with seamless A/B system updates according to Wikipedia is 7.1 and not 7.0

* Update .gitignore