Fix checkUtil method (#2)
* Add RootTools library

Added RootTools library to the project.

* Add RootShell library and exitApp function

Added the RootShell library which is the core of RootTools and moved the exit mechanism to a dedicated function.

* Change the code structure of A/B checking

Changed the code structure of the A/B checking and added support for checking Virtual A/B.

* Fix crashing when bootctl is missing

Fixed the crash when bootctl utility is missing.

* Root detection and integration of RootTools library

Added integration of RootTools and RootShell libraries - replaced all su executions with the libraries.
Added root detection dialog (if the device isn't rooted) and root denied dialog.

* Bottom layout margin

Added 10dp bottom margin in the layout.

* Changed version to v0.7

* Fixed get-suffix command

Moved the slot suffix command to the current slot command so it is created when the current slot is known.

* Update screenshot

Updated screenshot to v0.7

* Update README.md

Added credits.

* Activity lifecycle handling

Better handling of activity lifecycle. Added preservation of UI information and shell when activity get recreated. This allows snappy screen rotation and executing information commands only once.

* Fixed ViewModel integration

Assigning the shell to the ViewModel after it got created.

* Seperation of getting root shell

Seperation of getting a root shell to a dedicated function.

* Fixed back button exit

Fixed the bug that when the user presses the back button and opens the app again, there's a blank screen. The disadvantage with this fix is that the app immediately close with no closing animation.

* Move ABResult to ABChecker

Moved the ABResult to be inside ABChecker.

* Add bottom margin in landscape

Added the bottom margin also in the landscape layout.

* Add comments to some functions

* Modified checkUtil in RootTools library

Modified checkUtil in RootTools to accept also 750 permission.