Version 0.7 (#13)
* Merging V0.7 to master (#1)

* Add RootTools library
Added RootTools library to the project.

* Add RootShell library and exitApp function
Added the RootShell library which is the core of RootTools and moved the exit mechanism to a dedicated function.

* Change the code structure of A/B checking
Changed the code structure of the A/B checking and added support for checking Virtual A/B.

* Fix crashing when bootctl is missing
Fixed the crash when bootctl utility is missing.

* Root detection and integration of RootTools library
Added integration of RootTools and RootShell libraries - replaced all su executions with the libraries.
Added root detection dialog (if the device isn't rooted) and root denied dialog.

* Bottom layout margin
Added 10dp bottom margin in the layout.

* Changed version to v0.7

* Fixed get-suffix command
Moved the slot suffix command to the current slot command so it is created when the current slot is known.

* Update screenshot
Updated screenshot to v0.7
Added credits.

* Activity lifecycle handling
Better handling of activity lifecycle. Added preservation of UI information and shell when activity get recreated. This allows snappy screen rotation and executing information commands only once.

* Fixed ViewModel integration
Assigning the shell to the ViewModel after it got created.

* Separation of getting a root shell
Separation of getting a root shell to a dedicated function.

* Fixed back button exit
Fixed the bug that when the user presses the back button and opens the app again, there's a blank screen. The disadvantage with this fix is that the app immediately closes with no closing animation.

* Move ABResult to ABChecker
Moved the ABResult to be inside ABChecker.

* Add bottom margin in landscape
Added the bottom margin also in the landscape layout.

* Add comments to some functions

* Fix checkUtil method (#2)

* Update README.md

* Modified checkUtil in RootTools library

Modified checkUtil in RootTools to accept also 750 permission.

* Update settings.gradle
Removed duplication of include

* Update MainActivity.java
Updated a comment.

* Added constants for missing commands
Added constants for switch slot and reboot commands.

* Update comment
Updating the exitApp comment
Change SDK to SDK_INT instead of parsing the string.

* Fixed hanging shells
It now closing the shells but there are still hanging threads so I added calling to System.exit function.

* Change integration of libraries
Changed the integration of RootTools and RootShell libraries from pre-compiled to source code.

* Add checkUtilExecutable method
Added checkUtilExecutable to RootTools library and changed the call in MainActivity from checkUtil to the new method.

* Add new lines at the end of files
* Move closing of BufferedReader
Moved the closing of BufferedReader to before the return of the function.

* Remove su retries
Removed su retires when the access is denied.